+++ internal/lsp/range.go
@@ -35,16 +35,3 @@ func lspRangeToHCL(lspRng lsp.Range, f File) (*hcl.Range, error) {
 		End:      endPos,
 	}, nil
 }
-
-func HCLSeverityToLSP(severity hcl.DiagnosticSeverity) lsp.DiagnosticSeverity {
-	var sev lsp.DiagnosticSeverity
-	switch severity {
-	case hcl.DiagError:
-		sev = lsp.Error
-	case hcl.DiagWarning:
-		sev = lsp.Warning
-	case hcl.DiagInvalid:
-		panic("invalid diagnostic")
-	}
-	return sev
-}
