+++ internal/lsp/range_test.go
